[x]Best Practises Added [x]Dictation Fixed
