+++ app/Http/Controllers/IndexController.php
@@ -6,18 +6,18 @@ use Illuminate\Http\Request;
 
 class IndexController extends Controller
 {
-    public function index()
-    {
-        return view('welcome');
-    }
+    //public function index()
+    //{
+    //    return view('welcome');
+    //}
 
-    public function dashboard()
-    {
-        return view('user.dashboard');
-    }
-
-    public function profile()
-    {
-        return view('user.profile');
-    }
+    //public function dashboard()
+    //{
+    //    return view('user.dashboard');
+    //}
+    //
+    //public function profile()
+    //{
+    //    return view('user.profile');
+    //}
 }
