+++ app/Http/Livewire/Auth/Register.php
@@ -26,7 +26,7 @@ class Register extends Component
         ]);
 
         auth()->login($user);
-        return redirect(route('index'));
+        return redirect(route('home'));
 
     }
     public function render()
